requirements.txt
fastapi==0.115.11
httpx==0.28.1
logfire==3.8.0
pydantic==2.10.6
pydantic_ai==0.0.37
python-dotenv==1.0.1
uvicorn==0.34.0
